requirements.txt
Django===5.2.1
pillow
cloudinary
python-decouple
django-htmx